商品列表页完成

+++ public/front/js/search.js
@@ -34,16 +34,24 @@ $(function () {
   //删除单条记录
   // 1.注册点击事件
   $(".history").on( 'click' ,'i.mui-icon-closeempty', function () {
-    // 2.获取点击的历史记录的下标
-    var index = $(this).data("index");
+    // 增加确认框
+    mui.confirm( '你确认要删除这条历史记录吗?','温馨提示',['否','是'],function (e) {
+      var that = this;
+      if( e.index === 1 ){
+        // 2.获取点击的历史记录的下标
+        var index = $(that).data("index");
+          
+        var jsonStr = localStorage.getItem( 'search_List' );
+        var  arr = JSON.parse( jsonStr );
+        // 在数组里将对应项删除
+        arr.splice( index , 1  );
+        localStorage.setItem( 'search_List' , JSON.stringify( arr ));
+        //重新渲染
+        render();
+      }
+    } )
 
-    var jsonStr = localStorage.getItem( 'search_List' );
-    var  arr = JSON.parse( jsonStr );
-    // 在数组里将对应项删除
-    arr.splice( index , 1  );
-    localStorage.setItem( 'search_List' , JSON.stringify( arr ));
-    //重新渲染
-    render();
+    
 
   })
